feat: DAH-2931 simplify
